Feat: Add Subsidary Soft-Delete
소프트딜리트 기능 구현
유효 데이터를 한번 삭제 시 is_delete=True로 소프트 딜리트
리스트 목록에서는 is_delete=True인 데이터는 숨겨져서 표시

is_delete=True인 데이터를 한번 더 Delete 요청하면 영구적으로 삭제가 됩니다.

+++ restaurants/views.py
@@ -1,6 +1,8 @@
 from django.http import Http404
 from django.shortcuts import render
 
+from datetime                  import datetime,timedelta
+
 from rest_framework            import status
 from rest_framework.response   import Response
 from rest_framework.views      import APIView
@@ -16,8 +18,8 @@ class SubsidaryList(APIView):
     류성훈
     """
     def get(self, request, format=None):
-        # Subsidary의 모든 데이터를 읽어온다.
-        subsidary = Subsidary.objects.all()
+        # Subsidary의 모든 유효한 데이터를 읽어온다. 
+        subsidary = Subsidary.objects.filter(is_delete=False)
         serializer = SubsidarySerializer(subsidary, many=True)
         return Response(serializer.data)
 
@@ -53,8 +55,26 @@ class SubsidaryDetail(APIView):
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request, id, format=None):
-        subsidary = self.get_object(id)
-        subsidary.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+    # Soft Delete 방식 채택으로 인한 Delete기능 주석처리
+    # def delete(self, request, id, format=None):
+    #     subsidary = self.get_object(id)
+    #     subsidary.delete()
+    #     return Response(status=status.HTTP_204_NO_CONTENT)
 
+    # 한번 삭제 시 Soft Delete / Soft Delete 된 데이터 또 삭제 시 영구 Delete
+    def delete(self, request, id, format=None):
+        try:
+            subsidary = Subsidary.objects.get(id=id)
+            if subsidary.is_delete == True:
+                subsidary.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
+            serializer = SubsidarySerializer(subsidary)
+            subsidary.is_delete = True
+            subsidary.delete_at = datetime.now()
+            subsidary.save()
+            return Response(serializer, status=200)
+
+        except Subsidary.DoesNotExist:
+            return Response(status=404)
+
+
